Upgrade ethereumjs-lib to 3.0.0.
The current dependency on an old version results in a transitive dependency on secp256k1@1.1.4, which is broken on OSX. Newer versions of ethereumjs-lib resolve this issue (there are no newer minor revisions; the issue is only fixed in new major revisions).

Unit tests all still pass after this change.